Add composer.lock
We commit Website's composer.lock (along with composer.json) into
GitHub repository. Having it in the repository assures you that each
developer is using the same versions. This means also that the
deployment system always will install the expected versions of
dependencies. This is why after you pulled latest changes from the git
repository you have to run:

    composer install

instead of

    composer update

The complete command format is suitable for deployment in production:

    composer install \
        --prefer-dist \
        --no-interaction \
        --no-suggest \
        -o \
        --no-dev